Fix license, readme. Fix Coverage path processing on Windows and Linux Fix samples Make sure Coveralls in CI is actually doing its thing.
