+++ README.md
@@ -1,16 +1,13 @@
 PyBuilder
 =========
 
-[PyBuilder](http://pybuilder.github.io)
+[PyBuilder](https://pybuilder.github.io)
 
 
 [![Gitter](https://badges.gitter.im/pybuilder/pybuilder.svg)](https://gitter.im/pybuilder/pybuilder)
-[![Build Status](https://secure.travis-ci.org/pybuilder/pybuilder.png?branch=master)](http://travis-ci.org/pybuilder/pybuilder)
-[![Windows build status](https://ci.appveyor.com/api/projects/status/5jhel32oppeoqmw6/branch/master?svg=true)](https://ci.appveyor.com/project/arcivanov/pybuilder-yl8px/branch/master)
-[![PyPI version](https://badge.fury.io/py/pybuilder.png)](https://warehouse.python.org/project/pybuilder/)
+[![Build Status](https://secure.travis-ci.org/pybuilder/pybuilder.png?branch=master)](https://travis-ci.org/pybuilder/pybuilder)
+[![PyPI version](https://badge.fury.io/py/pybuilder.png)](https://pypi.org/project/pybuilder/)
 [![Coverage Status](https://coveralls.io/repos/pybuilder/pybuilder/badge.png?branch=master)](https://coveralls.io/r/pybuilder/pybuilder?branch=master)
-[![Ready in backlog](https://badge.waffle.io/pybuilder/pybuilder.png?label=ready&title=Ready)](https://waffle.io/pybuilder/pybuilder)
-[![Open bugs](https://badge.waffle.io/pybuilder/pybuilder.png?label=bug&title=Open%20Bugs)](https://waffle.io/pybuilder/pybuilder)
 
 PyBuilder is a software build tool written in 100% pure Python, mainly
 targeting Python applications.
@@ -19,7 +16,7 @@ PyBuilder is based on the concept of dependency based programming, but it also
 comes with a powerful plugin mechanism, allowing the construction of build life
 cycles similar to those known from other famous (Java) build tools.
 
-PyBuilder is running on the following versions of Python: 2.7, 3.4, 3.5, 3.6, 3.7, PyPy 2.7, and PyPy 3.5.
+PyBuilder is running on the following versions of Python: 2.7, 3.5, 3.6, 3.7, 3.8, and PyPy 2.7, 3.5 and 3.6.
 
 See the [Travis Build](https://travis-ci.org/pybuilder/pybuilder) for version specific output.
 
@@ -33,8 +30,7 @@ For development builds use:
 
     $ pip install --pre pybuilder
 
-See the [Cheeseshop page](https://warehouse.python.org/project/pybuilder/) for more
-information.
+See the [PyPI](https://pypi.org/project/pybuilder/) for more information.
 
 ## Getting started
 
@@ -53,13 +49,13 @@ use_plugin("python.distutils")
 default_task = "publish"
 ```
 
-See the [PyBuilder homepage](http://pybuilder.github.com/) for more details and
+See the [PyBuilder homepage](https://pybuilder.github.io) for more details and
 a list of plugins.
 
 ## Release Notes
 
-The release notes can be found [here](http://pybuilder.github.com/releasenotes/).
+The release notes can be found [here](https://pybuilder.github.com/releasenotes/).
 There will also be a git tag with each release. Please note that we do not currently promote tags to GitHub "releases".
 
 ## Development
-See [developing PyBuilder](http://pybuilder.github.io/documentation/developing_pybuilder.html)
+See [developing PyBuilder](https://pybuilder.github.io/documentation/developing_pybuilder.html)
