+++ samples/hello-world/build.py
@@ -15,9 +15,10 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
+
 from pybuilder.core import task
 
 
 @task
 def say_hello(logger):
-    logger.info("Hello, pybuilder")
+    logger.info("Hello, PyBuilder!")
