+++ samples/modern-python-app/build.py
@@ -15,10 +15,11 @@
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
+
 from pybuilder.core import use_plugin
 
 use_plugin("python.core")
-use_plugin("python.pyfix_unittest")
+use_plugin("python.unittest")
 use_plugin("python.coverage")
 use_plugin("python.distutils")
