--- samples/simple-python-app/src/unittest/python/helloworld_tests.py
@@ -1,30 +1,0 @@
-#   -*- coding: utf-8 -*-
-#
-#   This file is part of PyBuilder
-#
-#   Copyright 2011-2020 PyBuilder Team
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-import unittest
-
-from helloworld import helloworld
-from test_utils import Mock
-
-
-class HelloWorldTest(unittest.TestCase):
-    def test_should_issue_hello_world_message(self):
-        out = Mock()
-
-        helloworld(out)
-
-        out.write.assert_called_with("Hello world of Python\n")
